MPT-16437 Reorganise resource accounts mixins structure
