+++ mpt_api_client/resources/accounts/account.py
@@ -18,9 +18,11 @@ from mpt_api_client.resources.accounts.accounts_users import (
     AccountsUsersService,
     AsyncAccountsUsersService,
 )
-from mpt_api_client.resources.accounts.mixins import (
+from mpt_api_client.resources.accounts.mixins.activatable_mixin import (
     ActivatableMixin,
     AsyncActivatableMixin,
+)
+from mpt_api_client.resources.accounts.mixins.validate_mixin import (
     AsyncValidateMixin,
     ValidateMixin,
 )
