+++ mpt_api_client/resources/accounts/mixins/activatable_mixin.py
@@ -1,0 +1,55 @@
+from mpt_api_client.models import ResourceData
+
+
+class ActivatableMixin[Model]:
+    """Activatable mixin for activating, enabling, disabling and deactivating resources."""
+
+    def activate(self, resource_id: str, resource_data: ResourceData | None = None) -> Model:
+        """Activate a resource.
+
+        Args:
+            resource_id: Resource ID
+            resource_data: Resource data will be updated
+        """
+        return self._resource_action(  # type: ignore[attr-defined, no-any-return]
+            resource_id, "POST", "activate", json=resource_data
+        )
+
+    def deactivate(self, resource_id: str, resource_data: ResourceData | None = None) -> Model:
+        """Deactivate a resource.
+
+        Args:
+            resource_id: Resource ID
+            resource_data: Resource data will be updated
+        """
+        return self._resource_action(  # type: ignore[attr-defined, no-any-return]
+            resource_id, "POST", "deactivate", json=resource_data
+        )
+
+
+class AsyncActivatableMixin[Model]:
+    """Async activatable mixin for activating, enabling, disabling and deactivating resources."""
+
+    async def activate(self, resource_id: str, resource_data: ResourceData | None = None) -> Model:
+        """Activate a resource.
+
+        Args:
+            resource_id: Resource ID
+            resource_data: Resource data will be updated
+        """
+        return await self._resource_action(  # type: ignore[attr-defined, no-any-return]
+            resource_id, "POST", "activate", json=resource_data
+        )
+
+    async def deactivate(
+        self, resource_id: str, resource_data: ResourceData | None = None
+    ) -> Model:
+        """Deactivate a resource.
+
+        Args:
+            resource_id: Resource ID
+            resource_data: Resource data will be updated
+        """
+        return await self._resource_action(  # type: ignore[attr-defined, no-any-return]
+            resource_id, "POST", "deactivate", json=resource_data
+        )
