+++ mpt_api_client/resources/accounts/sellers.py
@@ -7,7 +7,10 @@ from mpt_api_client.http.mixins import (
 )
 from mpt_api_client.models import Model
 from mpt_api_client.models.model import ResourceData
-from mpt_api_client.resources.accounts.mixins import ActivatableMixin, AsyncActivatableMixin
+from mpt_api_client.resources.accounts.mixins.activatable_mixin import (
+    ActivatableMixin,
+    AsyncActivatableMixin,
+)
 
 
 class Seller(Model):
