+++ mpt_api_client/resources/accounts/users.py
@@ -11,7 +11,7 @@ from mpt_api_client.http.mixins import (
 )
 from mpt_api_client.models import Model
 from mpt_api_client.models.model import ResourceData
-from mpt_api_client.resources.accounts.mixins import (
+from mpt_api_client.resources.accounts.mixins.blockable_mixin import (
     AsyncBlockableMixin,
     BlockableMixin,
 )
