+++ tests/unit/resources/accounts/mixins/test_activatable_mixin.py
@@ -1,0 +1,158 @@
+import httpx
+import pytest
+import respx
+
+from mpt_api_client.http import AsyncService, Service
+from mpt_api_client.resources.accounts.mixins.activatable_mixin import (
+    ActivatableMixin,
+    AsyncActivatableMixin,
+)
+from tests.unit.conftest import DummyModel
+
+
+class DummyActivatableService(
+    ActivatableMixin[DummyModel],
+    Service[DummyModel],
+):
+    _endpoint = "/public/v1/dummy/activatable/"
+    _model_class = DummyModel
+    _collection_key = "data"
+
+
+class DummyAsyncActivatableService(
+    AsyncActivatableMixin[DummyModel],
+    AsyncService[DummyModel],
+):
+    _endpoint = "/public/v1/dummy/activatable/"
+    _model_class = DummyModel
+    _collection_key = "data"
+
+
+@pytest.fixture
+def activatable_service(http_client):
+    return DummyActivatableService(http_client=http_client)
+
+
+@pytest.fixture
+def async_activatable_service(async_http_client):
+    return DummyAsyncActivatableService(http_client=async_http_client)
+
+
+@pytest.mark.parametrize(
+    ("action", "input_status"),
+    [
+        ("activate", {"id": "OBJ-0000-0001", "status": "update"}),
+        ("deactivate", {"id": "OBJ-0000-0001", "status": "update"}),
+    ],
+)
+def test_activatable_resource_actions(activatable_service, action, input_status):
+    request_expected_content = b'{"id":"OBJ-0000-0001","status":"update"}'
+    response_expected_data = {"id": "OBJ-0000-0001", "status": "new_status"}
+    with respx.mock:
+        mock_route = respx.post(
+            f"https://api.example.com/public/v1/dummy/activatable/OBJ-0000-0001/{action}"
+        ).mock(
+            return_value=httpx.Response(
+                status_code=httpx.codes.OK,
+                headers={"content-type": "application/json"},
+                json=response_expected_data,
+            )
+        )
+
+        result = getattr(activatable_service, action)("OBJ-0000-0001", input_status)
+
+        assert mock_route.call_count == 1
+        request = mock_route.calls[0].request
+        assert request.content == request_expected_content
+        assert result.to_dict() == response_expected_data
+        assert isinstance(result, DummyModel)
+
+
+@pytest.mark.parametrize(
+    ("action", "input_status"),
+    [
+        ("activate", None),
+        ("deactivate", None),
+    ],
+)
+def test_actions_no_data(activatable_service, action, input_status):
+    request_expected_content = b""
+    response_expected_data = {"id": "OBJ-0000-0001", "status": "new_status"}
+    with respx.mock:
+        mock_route = respx.post(
+            f"https://api.example.com/public/v1/dummy/activatable/OBJ-0000-0001/{action}"
+        ).mock(
+            return_value=httpx.Response(
+                status_code=httpx.codes.OK,
+                headers={"content-type": "application/json"},
+                json=response_expected_data,
+            )
+        )
+
+        result = getattr(activatable_service, action)("OBJ-0000-0001", input_status)
+
+        assert mock_route.call_count == 1
+        request = mock_route.calls[0].request
+        assert request.content == request_expected_content
+        assert result.to_dict() == response_expected_data
+        assert isinstance(result, DummyModel)
+
+
+@pytest.mark.parametrize(
+    ("action", "input_status"),
+    [
+        ("activate", {"id": "OBJ-0000-0001", "status": "update"}),
+        ("deactivate", {"id": "OBJ-0000-0001", "status": "update"}),
+    ],
+)
+async def test_async_activatable_resource_actions(async_activatable_service, action, input_status):
+    request_expected_content = b'{"id":"OBJ-0000-0001","status":"update"}'
+    response_expected_data = {"id": "OBJ-0000-0001", "status": "new_status"}
+    with respx.mock:
+        mock_route = respx.post(
+            f"https://api.example.com/public/v1/dummy/activatable/OBJ-0000-0001/{action}"
+        ).mock(
+            return_value=httpx.Response(
+                status_code=httpx.codes.OK,
+                headers={"content-type": "application/json"},
+                json=response_expected_data,
+            )
+        )
+
+        result = await getattr(async_activatable_service, action)("OBJ-0000-0001", input_status)
+
+        assert mock_route.call_count == 1
+        request = mock_route.calls[0].request
+        assert request.content == request_expected_content
+        assert result.to_dict() == response_expected_data
+        assert isinstance(result, DummyModel)
+
+
+@pytest.mark.parametrize(
+    ("action", "input_status"),
+    [
+        ("activate", None),
+        ("deactivate", None),
+    ],
+)
+async def test_async_actions_no_data(async_activatable_service, action, input_status):
+    request_expected_content = b""
+    response_expected_data = {"id": "OBJ-0000-0001", "status": "new_status"}
+    with respx.mock:
+        mock_route = respx.post(
+            f"https://api.example.com/public/v1/dummy/activatable/OBJ-0000-0001/{action}"
+        ).mock(
+            return_value=httpx.Response(
+                status_code=httpx.codes.OK,
+                headers={"content-type": "application/json"},
+                json=response_expected_data,
+            )
+        )
+
+        result = await getattr(async_activatable_service, action)("OBJ-0000-0001", input_status)
+
+        assert mock_route.call_count == 1
+        request = mock_route.calls[0].request
+        assert request.content == request_expected_content
+        assert result.to_dict() == response_expected_data
+        assert isinstance(result, DummyModel)
